Added display list for event, fixed date view and fixed pay button amount
